Start adding tests
