+++ source/Octopus.Tentacle/Kubernetes/KubernetesPodMonitor.cs
@@ -118,112 +118,6 @@ namespace Octopus.Tentacle.Kubernetes
             => podStatusLookup.TryGetValue(scriptTicket, out var status) ? status : null;
     }
 
-    public class ScriptPodStatus
-    {
-        readonly IKubernetesPodService podService;
-        Task? backgroundTask;
-        CancellationTokenSource? cancellationTokenSource;
-
-        List<LogLine> logLines = new();
-
-        static readonly object LogLock = new();
-        readonly string podName;
-        readonly string containerName;
-
-        long currentSequenceNumber = 0;
-
-        public ScriptPodStatus(V1Pod pod, IKubernetesPodService podService)
-        {
-            this.podService = podService;
-            podName = pod.Name();
-            containerName = pod.Spec.Containers.First().Name;
-        }
-
-        public void StartMonitoring(CancellationToken cancellationToken)
-        {
-            cancellationTokenSource = CancellationTokenSource.CreateLinkedTokenSource(cancellationToken);
-
-            backgroundTask = Task.Run(() => WatchPodLogsAsync(cancellationTokenSource.Token), cancellationToken);
-        }
-
-        public void StopMonitoring()
-        {
-            if (backgroundTask is null || cancellationTokenSource is null)
-                return;
-
-            cancellationTokenSource.Cancel();
-            backgroundTask.Wait();
-        }
-
-        public (long newSequence, List<LogLine>) GetLogLines(long lastLogSequence)
-        {
-            lock (LogLock)
-            {
-                // we determine how many lines to retrieve
-                // this number is how many tail lines we need to return
-                var linesToGet = (logLines.Count + currentSequenceNumber) - lastLogSequence;
-
-                var newSequence = lastLogSequence + linesToGet;
-
-                // shortcut if we are return the entire list
-                var linesToReturn = linesToGet == logLines.Count
-                    ? logLines.ToList()
-                    : TakeLast(logLines, linesToGet).ToList();
-
-                //clear the current log lines list
-                logLines.Clear();
-                //update our sequence number
-                currentSequenceNumber = newSequence;
-
-                return (newSequence, linesToReturn);
-
-                // Because we are mixing longs and int's, this becomes the easiest way to do this.
-                IEnumerable<LogLine> TakeLast(IReadOnlyList<LogLine> lines, long count)
-                {
-                    var startingIndex = lines.Count - count;
-                    for (var i = 0; i < lines.Count; i++)
-                    {
-                        if (i <= startingIndex)
-                        {
-                            continue;
-                        }
-
-                        yield return lines[i];
-                    }
-                }
-            }
-        }
-
-        async Task WatchPodLogsAsync(CancellationToken cancellationToken)
-        {
-            await foreach (var logLine in podService.StreamPodLogs(podName, containerName, cancellationToken))
-            {
-                if (logLine is null)
-                    continue;
-
-                var parts = logLine.Split('|');
-                if (parts.Length != 3)
-                    continue;
-
-                var occured = DateTimeOffset.Parse(parts[0]);
-                var source = parts[1] switch
-                {
-                    "stdout" => ProcessOutputSource.StdOut,
-                    "stderr" => ProcessOutputSource.StdErr,
-                    _ => throw new InvalidOperationException($"Unknown source {parts[1]}")
-                };
-
-                lock (LogLock)
-                {
-                    //we are making a bold assumption that the pod logs are coming in sequential order
-                    logLines.Add(new LogLine(occured, source, parts[2]));
-                }
-            }
-        }
-
-        public record LogLine(DateTimeOffset Occurred, ProcessOutputSource Source, string Message);
-    }
-
     public class PodStatus
     {
         public ScriptTicket ScriptTicket { get; }
